Pulled header out of App.tsx to allow for page-specific content. Slight refactoring of components directory.
