+++ src/App.tsx
@@ -1,6 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import { BrowserRouter as Router, Link, Route, Switch } from 'react-router-dom';
-import{ Button, Pane, majorScale } from 'evergreen-ui';
+import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 
 import Add from './pages/Add';
 import List from './pages/List';
@@ -34,22 +33,17 @@ const App = () => {
   return (
     <Router>
       <TagContext.Provider value={dedupe(recipes.reduce((array, { tags }) => [...array, ...tags], []))}>
-        <Pane background="tealTint" padding={majorScale(3)}>
-          <Button is={Link} to="/add">Add Recipe</Button>
-        </Pane>
-        <Pane padding={majorScale(3)}>
-          <Switch>
-            <Route path="/add">
-              <Add user={user} recipes={recipes} />
-            </Route>
-            <Route path="/view/:id">
-              <Recipe user={user} recipes={recipes} />
-            </Route>
-            <Route path="*">
-              <List recipes={recipes} />
-            </Route>
-          </Switch>
-        </Pane>
+        <Switch>
+          <Route path="/add">
+            <Add user={user} recipes={recipes} />
+          </Route>
+          <Route path="/view/:id">
+            <Recipe user={user} recipes={recipes} />
+          </Route>
+          <Route path="*">
+            <List recipes={recipes} />
+          </Route>
+        </Switch>
       </TagContext.Provider>
     </Router>
   );
