+++ src/components/index.ts
@@ -1,0 +1,3 @@
+export { default as Layout } from './Layout';
+export { default as ListInput } from './ListInput';
+export { default as TagInput } from './TagInput';
